# Performance Optimisation Around Dependency Calculations