… go/vt/vtgate/planbuilder/testdata/my.json