```
                                            │ ../old.bench │            ../new.bench             │
                                            │    sec/op    │   sec/op     vs base                │
OLTP-20                                        238.5µ ± 6%   252.7µ ± 2%   +5.94% (p=0.005 n=10)
TPCC-20                                        1.702m ± 3%   1.760m ± 1%   +3.42% (p=0.000 n=10)
TPCH-20                                        10.08m ± 1%   10.06m ± 1%        ~ (p=0.971 n=10)
Planner/from_cases.json-gen4-20                7.376m ± 1%   7.502m ± 1%   +1.70% (p=0.000 n=10)
Planner/filter_cases.json-gen4-20              147.5m ± 0%   149.4m ± 0%   +1.27% (p=0.000 n=10)
Planner/large_cases.json-gen4-20               464.7µ ± 0%   464.5µ ± 1%        ~ (p=0.631 n=10)
Planner/aggr_cases.json-gen4-20                13.10m ± 0%   13.13m ± 0%        ~ (p=0.315 n=10)
Planner/select_cases.json-gen4-20              10.66m ± 1%   10.65m ± 1%        ~ (p=0.971 n=10)
Planner/union_cases.json-gen4-20               3.808m ± 0%   3.879m ± 0%   +1.87% (p=0.000 n=10)
SemAnalysis-20                                 28.74m ± 1%   29.26m ± 1%   +1.81% (p=0.000 n=10)
SelectVsDML/DML_(random_sample,_N=32)-20       875.9µ ± 1%   965.0µ ± 2%  +10.17% (p=0.000 n=10)
SelectVsDML/Select_(random_sample,_N=32)-20    1.871m ± 0%   1.622m ± 1%  -13.30% (p=0.000 n=10)
BaselineVsMirrored/Baseline-20                 126.2µ ± 1%   128.4µ ± 1%   +1.71% (p=0.000 n=10)
BaselineVsMirrored/Mirrored-20                 333.7µ ± 0%   340.2µ ± 1%   +1.95% (p=0.000 n=10)
Mine-20                                        5.684m ± 0%   5.409m ± 0%   -4.84% (p=0.000 n=10)
geomean                                        2.971m        2.990m        +0.65%

                                            │ ../old.bench  │             ../new.bench             │
                                            │     B/op      │     B/op      vs base                │
OLTP-20                                        155.1Ki ± 0%   155.3Ki ± 0%   +0.15% (p=0.000 n=10)
TPCC-20                                        1.029Mi ± 0%   1.030Mi ± 0%   +0.11% (p=0.000 n=10)
TPCH-20                                        5.324Mi ± 0%   5.135Mi ± 0%   -3.55% (p=0.000 n=10)
Planner/from_cases.json-gen4-20                4.567Mi ± 0%   4.550Mi ± 0%   -0.37% (p=0.000 n=10)
Planner/filter_cases.json-gen4-20              20.13Mi ± 0%   20.05Mi ± 0%   -0.41% (p=0.000 n=10)
Planner/large_cases.json-gen4-20               276.6Ki ± 0%   277.4Ki ± 0%   +0.29% (p=0.000 n=10)
Planner/aggr_cases.json-gen4-20                7.350Mi ± 0%   7.196Mi ± 0%   -2.10% (p=0.000 n=10)
Planner/select_cases.json-gen4-20              6.334Mi ± 0%   6.203Mi ± 0%   -2.08% (p=0.000 n=10)
Planner/union_cases.json-gen4-20               2.269Mi ± 0%   2.272Mi ± 0%   +0.14% (p=0.000 n=10)
SelectVsDML/DML_(random_sample,_N=32)-20       594.6Ki ± 0%   652.6Ki ± 0%   +9.76% (p=0.000 n=10)
SelectVsDML/Select_(random_sample,_N=32)-20   1128.1Ki ± 0%   975.8Ki ± 0%  -13.50% (p=0.000 n=10)
BaselineVsMirrored/Baseline-20                 88.03Ki ± 0%   88.04Ki ± 0%   +0.01% (p=0.033 n=10)
BaselineVsMirrored/Mirrored-20                 208.0Ki ± 0%   208.3Ki ± 0%   +0.13% (p=0.000 n=10)
Mine-20                                        3.515Mi ± 0%   3.243Mi ± 0%   -7.76% (p=0.000 n=10)
geomean                                        1.374Mi        1.353Mi        -1.49%

                                            │ ../old.bench │             ../new.bench             │
                                            │  allocs/op   │  allocs/op   vs base                 │
OLTP-20                                        3.859k ± 0%   3.929k ± 0%  +1.81% (p=0.000 n=10)
TPCC-20                                        25.63k ± 0%   26.02k ± 0%  +1.55% (p=0.000 n=10)
TPCH-20                                        134.9k ± 0%   136.1k ± 0%  +0.88% (p=0.000 n=10)
Planner/from_cases.json-gen4-20                107.9k ± 0%   110.6k ± 0%  +2.48% (p=0.000 n=10)
Planner/filter_cases.json-gen4-20              534.4k ± 0%   538.1k ± 0%  +0.68% (p=0.000 n=10)
Planner/large_cases.json-gen4-20               10.71k ± 0%   11.29k ± 0%  +5.42% (p=0.000 n=10)
Planner/aggr_cases.json-gen4-20                168.1k ± 0%   171.3k ± 0%  +1.89% (p=0.000 n=10)
Planner/select_cases.json-gen4-20              147.8k ± 0%   150.7k ± 0%  +1.95% (p=0.000 n=10)
Planner/union_cases.json-gen4-20               53.22k ± 0%   54.11k ± 0%  +1.67% (p=0.000 n=10)
SelectVsDML/DML_(random_sample,_N=32)-20       12.53k ± 0%   13.20k ± 0%  +5.36% (p=0.000 n=10)
SelectVsDML/Select_(random_sample,_N=32)-20    25.82k ± 0%   23.52k ± 0%  -8.93% (p=0.000 n=10)
BaselineVsMirrored/Baseline-20                 2.276k ± 0%   2.276k ± 0%       ~ (p=1.000 n=10) ¹
BaselineVsMirrored/Mirrored-20                 5.837k ± 0%   6.006k ± 0%  +2.90% (p=0.000 n=10)
Mine-20                                        74.51k ± 0%   72.65k ± 0%  -2.50% (p=0.000 n=10)
geomean                                        34.58k        34.93k       +1.02%
```

<!-- if this PR is Work in Progress please create it as a Draft Pull Request -->

## Description

<!-- A few sentences describing the overall goals of the pull request's commits. -->
<!-- If this is a bug fix and you think the fix should be backported, please write so. -->

## Related Issue(s)

<!-- List related issues and pull requests. If this PR fixes an issue, please add it using Fixes #????  -->

## Checklist

-   [ ] "Backport to:" labels have been added if this change should be back-ported to release branches
-   [ ] If this change is to be back-ported to previous releases, a justification is included in the PR description
-   [ ] Tests were added or are not required
-   [ ] Did the new or modified tests pass consistently locally and on CI?
-   [ ] Documentation was added or is not required

## Deployment Notes

<!-- Notes regarding deployment of the contained body of work. These should note any db migrations, etc. -->